upload

+++ models/contacts.js
@@ -1,17 +1,17 @@
 const Contact = require('../service/schemas/schemaContacts');
 
-const listContacts = async () => {
+const listContacts = async userId => {
   try {
-    return Contact.find();
+    return Contact.find({ owner: userId });
   } catch (error) {
     console.error('Error occurred when trying to show contacts:', error);
     throw error;
   }
 };
 
-const getContactById = async contactId => {
+const getContactById = async (contactId, userId) => {
   try {
-    const contact = await Contact.findById(contactId);
+    const contact = await Contact.findById({ _id: contactId, owner: userId });
     return contact;
   } catch (error) {
     console.error('Error occurred when trying to get contact:', error);
@@ -19,9 +19,9 @@ const getContactById = async contactId => {
   }
 };
 
-const removeContact = async contactId => {
+const removeContact = async (contactId, userId) => {
   try {
-    const result = await Contact.findByIdAndRemove(contactId);
+    const result = await Contact.findByIdAndRemove({ _id: contactId, owner: userId });
     return result;
   } catch (error) {
     console.error('Error occurred when removing contact:', error);
@@ -29,9 +29,9 @@ const removeContact = async contactId => {
   }
 };
 
-const addContact = async body => {
+const addContact = async (body, userId) => {
   try {
-    const newContact = await Contact.create(body);
+    const newContact = await Contact.create({ ...body, owner: userId });
     return newContact;
   } catch (error) {
     console.error('Error occurred when adding contact:', error);
@@ -39,9 +39,13 @@ const addContact = async body => {
   }
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, userId) => {
   try {
-    const updatedContact = await Contact.findByIdAndUpdate(contactId, body, { new: true });
+    const updatedContact = await Contact.findByIdAndUpdate(
+      { _id: contactId, owner: userId },
+      body,
+      { new: true }
+    );
     return updatedContact;
   } catch (error) {
     console.error('Error occurred when updating contact:', error);
@@ -49,9 +53,14 @@ const updateContact = async (contactId, body) => {
   }
 };
 
-const updateFavorite = async (contactId, favorite) => {
+const updateFavorite = async (contactId, favorite, userId) => {
   try {
-    const updatedContact = await Contact.findByIdAndUpdate(contactId, { favorite }, { new: true });
+    const updatedContact = await Contact.findOneAndUpdate(
+      { _id: contactId, owner: userId },
+      { favorite },
+      { new: true }
+    );
+
     return updatedContact;
   } catch (error) {
     console.error('Error occurred when updating contact:', error);
